mmc: core: Enable HS mode first in the device and then in the host
While switching to HS mode, high speed mode timing should be selected
in the device before changing the clock frequency in the host.

But present driver first updates the frequency in host and then selects
HS mode in device while selecting HS400 mode. This is a spec violation.
Updated the sequence to comply with spec.

Change-Id: I5b2c1f724d820daf9c0bca8651cf85bd0ff67655
Signed-off-by: Veerabhadrarao Badiganti <vbadigan@codeaurora.org>